Run spotlessApply
