--- shared-ios/src/iosMain/kotlin/main.ios.kt
@@ -1,6 +1,0 @@
-import androidx.compose.ui.window.ComposeUIViewController
-import soup.compose.material.motion.shared.ui.NavGraph
-
-fun MainViewController() = ComposeUIViewController {
-    NavGraph()
-}
